server/requirements.txt
flask==3.0.2
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
yt-dlp==2024.3.10
instaloader==4.10.3
python-telegram-bot==20.8
bcrypt==4.1.2
PyJWT==2.8.0
requests==2.31.0